@@ -1,24 +1,25 @@
 module github.com/homedepot/go-clouddriver
 
-go 1.18
+go 1.21
 
 require (
 	cloud.google.com/go/storage v1.17.0
 	github.com/DATA-DOG/go-sqlmock v1.5.0
 	github.com/fatih/color v1.9.0
 	github.com/fsouza/fake-gcs-server v1.30.1
-	github.com/gin-gonic/gin v1.6.3
+	github.com/gin-gonic/gin v1.9.1
 	github.com/google/go-github/v32 v32.1.0
 	github.com/google/uuid v1.1.2
 	github.com/googleapis/gnostic v0.5.1
 	github.com/gregjones/httpcache v0.0.0-20180305231024-9cad4c3443a7
+	github.com/homedepot/arcade v1.3.0
 	github.com/iancoleman/strcase v0.1.2
 	github.com/jonboulle/clockwork v0.1.0
-	github.com/onsi/ginkgo/v2 v2.1.4
-	github.com/onsi/gomega v1.19.0
+	github.com/onsi/ginkgo/v2 v2.13.0
+	github.com/onsi/gomega v1.30.0
 	github.com/peterbourgon/diskv v2.0.1+incompatible
 	github.com/zsais/go-gin-prometheus v0.1.1-0.20200217150448-2199a42d96c1
-	golang.org/x/oauth2 v0.0.0-20210819190943-2bc19b11175f
+	golang.org/x/oauth2 v0.15.0
 	google.golang.org/api v0.57.0
 	gopkg.in/yaml.v2 v2.4.0
 	gorm.io/driver/mysql v1.3.4
