@@ -3,7 +3,7 @@ package fiat
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 )
 
@@ -80,7 +80,7 @@ func (c *client) Authorize(account string) (Response, error) {
 		return Response{}, fmt.Errorf("user authorization error: %s", res.Status)
 	}
 
-	b, err := ioutil.ReadAll(res.Body)
+	b, err := io.ReadAll(res.Body)
 	if err != nil {
 		return Response{}, err
 	}
