@@ -18,7 +18,7 @@ package memory_test
 
 import (
 	"bytes"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/url"
 
@@ -77,7 +77,7 @@ var _ = Describe("CachedDiscovery", func() {
 		JustBeforeEach(func() {
 			resp, err = cache.RoundTrip(req)
 			Expect(err).To(BeNil())
-			content, err = ioutil.ReadAll(resp.Body)
+			content, err = io.ReadAll(resp.Body)
 			Expect(err).To(BeNil())
 		})
 
@@ -85,7 +85,7 @@ var _ = Describe("CachedDiscovery", func() {
 			BeforeEach(func() {
 				rt.Response = &http.Response{
 					Header:     http.Header{"ETag": []string{`"123456"`}},
-					Body:       ioutil.NopCloser(bytes.NewReader([]byte("Content"))),
+					Body:       io.NopCloser(bytes.NewReader([]byte("Content"))),
 					StatusCode: http.StatusOK,
 				}
 			})
@@ -99,7 +99,7 @@ var _ = Describe("CachedDiscovery", func() {
 			BeforeEach(func() {
 				rt.Response = &http.Response{
 					StatusCode: http.StatusNotModified,
-					Body:       ioutil.NopCloser(bytes.NewReader([]byte("Other Content"))),
+					Body:       io.NopCloser(bytes.NewReader([]byte("Other Content"))),
 				}
 			})
 
